Fix crash caused by days with empty subst table
Days without substitutions result in a table that consists of only one
row. This row contains a single cell with the text "Keine Vertretungen".
The parser assumed that the first row of the table is the header and
hence failed to parse this empty table. This is now fixed by introducing
an additional check for single-row tables.